backend/requirements.txt
cryptography
Flask==3.1.2
flask-cors==6.0.1
flask-sqlalchemy
flask-jwt-extended
pymysql
requests
joblib
scikit-learn
Pillow
numpy
opencv-python
tensorflow==2.15.0
googlesearch-python
beautifulsoup4
blinker==1.9.0
click==8.3.0
colorama==0.4.6
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.3